Devitalization of gui.glade (now renamed in gui/gui.xml) and major reorganization of gui-related code The word "devitalization" means that the glade generated file doesn't contain any translatable string. Such strings are now in gui/<source>.ml and are thus ready to be encapsulated in a call for translation. The sub-directory gui/ is now dedicated to sources completing the glade specification. This "completion" concerns:  1) translatable strings  2) dynamic widgets
Names of gui/ files follow a convention: relevant widget defined in glade are named using capital letters.
If a source file is created in order to complete a widget, it takes in gui/ the name of the widget prefixed 
by "gui_" and suffixed by ".ml". For instance, the widget menubar_FOO will be implemented by 
gui/gui_menubar_FOO.ml.

Also, the gui-related code has been massively refactored. The file talking.ml still exists but it has been
split into several sources, usually 2 per component. Actually, each component "foo" with a dialog 
"dialog_FOO" defined with glade, has essentially two related files:
 1) gui/gui_foo.ml
 2) gui/gui_dialog_FOO.ml
The first file defines callbacks, while the second encapsulated the glade dialog definition
in order to build a dialog returning an environment, i.e. a dialog with a more abstract behaviour. 
Component-related menus are created with a couple of functors defined in:  
./gui/gui_toolbar_COMPONENTS_layouts.ml{,i}

Two templates:
./gui/gui_component-node-with-state.ml-template
./gui/gui_component-node-without-state.ml-template
show how to create the gui-related part of a new component. The first show how to implement
the gui for components without state (like an hub), while the second show how to implement the gui 
for a components with state (like a machine, i.e. with a ungracefully_stop action).