Added support for multi-threading in the chip/wire/system machinery The basic get/set operations on wires are now controled with simple mutex. Regarding source chips, the emit operation is also performed in thread-mutual exclusion.
Luca's recursive mutex are not necessary in this version even if they could be useful 
for preventing deadlocks in the case of a bad designed system (motherboard).
For "bad designed" I mean a system with a chip-stabilization function calling 
(directly or not) a get/set on a wire not *explicitely* connected to its ports.

The get method of chips (generated by the camlp4 extension) call now the methods get_alone 
of its output ports. Thus, the name "get_alone" for this method seems to be more appropriate but, 
in the same time, not so user-mnemonic as "get". Also the name "get_alone" seems to me not very
self-explanatory: other possibilities could be "unsafe_get" or "private_get". 
I will take a decision about names in the future.